Bug 32898: Avoid hardcoding user/pass in tests
This patch makes the authentication parameters be configurable through
env. Before this, it was hardcoded to koha/koha, which only works if the
ktd defaults haven't been tweaked.

If no env is defined, it fallsback to koha/koha.

In order to pass the ktd configured params, the tests need to be run
like:

$ yarn cypress run --env KOHA_USER=$KOHA_USER,KOHA_PASS=$KOHA_PASS

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 53bebb413aec3796ce66311b0e66c872e01fee19)
Signed-off-by: Jacob O'Mara <jacob.omara@ptfs-europe.com>