Bug 32752: Update serial edit form to allow for new status
With this patch the new status Circulating, Out for binding and
Bound can be set for any issue, if the current status is not "Expected".
This is to keep the list a little shorter when the issue hasn't even
arrived in the library yet.

These new status will be especially helpful if the library has
opted to not create items for all their subscriptions as a lot
of libraries will only lend bound volumes, but not the singular
issues.

* Circulating: this will be useful if the library is using
routing lists and the items are not available at the library
for a longer time.

* Out for binding: if you send out issues for binding this will
help to track these.

* Bound: this will help users to know that the issue is now
available as a bound volume.

To test:
* If you haven't yet:
  * Create a subscription
  * From the serial collection page:
    * Receive or create several issues and set different status
    * Make sure you have at least 2 late and claim one of them
    * Make sure you have several Arrived issues
  * Edit Arrived serial issues and others
  * Verify that the status pull down no also shows the new status
    if the issue has a different status than "Expected"
  * Set the new status
  * Verify they save and show correctly
  * Edit again and verify the pull downs are pre-set correctly
    to the saved status
  * Verify they show everywhere else (see previous test plans :) )

Signed-off-by: Michaela Sieber <michaela.sieber@kit.edu>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>